rename analytics/svc -> analytics/service
