+++ cmd/state-svc/service.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	"github.com/ActiveState/cli/cmd/state-svc/internal/server"
-	"github.com/ActiveState/cli/internal/analytics/svc"
+	anaSvc "github.com/ActiveState/cli/internal/analytics/service"
 	"github.com/ActiveState/cli/internal/config"
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
@@ -17,12 +17,12 @@ import (
 
 type service struct {
 	cfg      *config.Instance
-	an       *svc.Analytics
+	an       *anaSvc.Analytics
 	shutdown context.CancelFunc
 	server   *server.Server
 }
 
-func NewService(cfg *config.Instance, an *svc.Analytics, shutdown context.CancelFunc) *service {
+func NewService(cfg *config.Instance, an *anaSvc.Analytics, shutdown context.CancelFunc) *service {
 	return &service{cfg: cfg, an: an, shutdown: shutdown}
 }
